[logic][m]: delete data package method moved into logic layer. tests updated imports updated, mocks updated - refs #401
