+++ app/logic/db_logic.py
@@ -108,3 +108,20 @@ def change_package_status(publisher_name, package_name, status=PackageStateEnum.
     db.session.add(data)
     db.session.commit()
     return True
+
+
+def delete_data_package(publisher_name, package_name):
+    """
+    This method deletes the data package. This method used
+    for hard delete the data package
+    :param publisher_name: publisher name
+    :param package_name: package name
+    :return: If success True else False
+    """
+    data = Package.query.join(Publisher). \
+        filter(Publisher.name == publisher_name,
+               Package.name == package_name).one()
+    package_id = data.id
+    Package.query.filter(Package.id == package_id).delete()
+    db.session.commit()
+    return True
